+++ src/components/Form/Form.component.tsx
@@ -26,6 +26,7 @@ const defaultFormProps: FormDefaultProps<object> = {
   onChange: null,
   onSubmit: null,
   onSubmitError: console.error,
+  onCancelled: null,
   title: "Form",
   badge: null,
   icon: null,
@@ -38,6 +39,8 @@ const defaultFormProps: FormDefaultProps<object> = {
   debug: false,
   asModal: null,
   columns: 2,
+  submitLabel: "save",
+  cancelLabel: "cancel",
 };
 
 /**
@@ -65,6 +68,10 @@ const Form = <TData extends object>(
   );
 };
 
+Form.defaultProps = defaultFormProps;
+
+export { Form };
+
 /**
  * Renders the inner form component with specific elements.
  *
@@ -85,7 +92,10 @@ const InnerForm = <TData extends object>(
       <FormHeader {...props} />
       <FormFieldRenderer />
       <FormDebug />
-      <FormEditBanner />
+      <FormEditBanner
+        submitLabel={props.submitLabel}
+        cancelLabel={props.cancelLabel}
+      />
     </FormContainer>
   );
 };
@@ -93,6 +103,3 @@ const InnerForm = <TData extends object>(
 const FormContainer = styled.form<FormContainerStyleProps>`
   ${formContainerStyles}
 `;
-Form.defaultProps = defaultFormProps;
-
-export { Form };