+++ src/utils/translation/translation.keys.ts
@@ -33,6 +33,7 @@ const actionTranslationKeys = [
   "collapse",
   "show",
   "hide",
+  "giveUp",
 ] as const;
 
 const validationTranslationKeys = [